update all method
