+++ lib/board.rb
@@ -1,0 +1,5 @@
+module Board
+  def initialize
+    @board = [' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ', ' ']
+  end
+end
